Restructure the reducers so that we can use separate functions
